[SPARK-46451][PS][TESTS] Reorganize `GroupbyStatTests`
### What changes were proposed in this pull request?
Reorganize `GroupbyStatTests`

### Why are the changes needed?
`GroupbyStatTests` and its parity test are slow, factor the slow tests out to make them more suiable for parallelism

### Does this PR introduce _any_ user-facing change?
no, test-only

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44409 from zhengruifeng/ps_test_group_stat.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
